Isolate legacy NinSnes implementation
