+++ src/main/formats/NinSnes/legacy/NinSnesLegacyScanner.cpp
@@ -4,8 +4,8 @@
  * refer to the included LICENSE.txt file
  */
 
-#include "NinSnesInstr.h"
-#include "NinSnesSeq.h"
+#include "../NinSnesInstr.h"
+#include "../NinSnesSeq.h"
 #include "ScannerManager.h"
 
 namespace vgmtrans::scanners {
